% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\docType{data}
\name{credentials}
\alias{credentials}
\title{credentials data}
\format{
2 rows, 7 columns.
}
\usage{
data("credentials")
}
\description{
This data is the credentials dataset.
}
\examples{
data("credentials")

}
\keyword{datasets}
