% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{AFD open data}
\format{
1820 rows, 37 columns.
}
\usage{
data("dataset")
}
\description{
This data set is a small example of AFD dataset.
}
\examples{
data("dataset")

}
\keyword{datasets}
